requirements.txt
# PDF processing
PyMuPDF==1.22.5  # For extracting text from PDF files

# OpenAI API interaction
openai==1.3.5  # For interacting with OpenAI's language models

# Slack integration
slack_sdk==3.26.0  # For posting messages to Slack

# Configuration management
PyYAML==6.0.1  # For reading YAML configuration files

# Testing
pytest==7.4.3  # For running unit tests

# Environment variable management
python-dotenv==1.0.0  # For loading environment variables from .env file

# Logging (part of Python standard library, no installation needed)
# import logging